19. gün 3. ödev

+++ ReCapProject.Business/Concrete/RentalEngine.cs
@@ -1,11 +1,13 @@
 ﻿using ReCapProject.Business.Abstract;
 using ReCapProject.Business.Constants;
 using ReCapProject.Business.ValidationRules.FluentValidation;
+using ReCapProject.Core.Aspects.Autofac.Transaction;
 using ReCapProject.Core.Aspects.Autofac.Validation;
 using ReCapProject.Core.Utilities.Results;
 using ReCapProject.Data.Access.Abstract;
 using ReCapProject.Data.Entities;
 using ReCapProject.Data.Entities.DTOs;
+using ReCapProject.Data.Entities.ResponseModels;
 using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
@@ -15,14 +17,23 @@ namespace ReCapProject.Business.Concrete
     public class RentalEngine : IRentalEngine
     {
         private readonly IRentalRepository _rentalRepository;
+        private readonly IPaymentEngine _paymentEngine;
 
-        public RentalEngine(IRentalRepository rentalRepository)
+        public RentalEngine(IRentalRepository rentalRepository, IPaymentEngine paymentEngine)
         {
             _rentalRepository = rentalRepository;
+            _paymentEngine = paymentEngine;
         }
 
         public IResult Delete(Rental rental)
         {
+            var mapResult = _paymentEngine.GetMaps(map => map.RentalId == rental.Id);
+            if (!mapResult.Success)
+            {
+                return mapResult;
+            }
+            (mapResult.Data as List<PaymentRentalMap>).ForEach(map => _paymentEngine.DeleteMap(map));
+
             _rentalRepository.Delete(rental);
             _rentalRepository.SaveChanges();
             return new SuccessResult();
@@ -78,6 +89,48 @@ namespace ReCapProject.Business.Concrete
             {
                 return new ErrorDataResult<Rental>(Messages.CarIsRented);
             }
+        }
+
+        [TransactionScopeAspect]
+        public IDataResult<Rental> InsertWithPayment(RentalResponseModel rentalResponseModel)
+        {
+            var payment = new Payment()
+            {
+                CardHolderName = rentalResponseModel.CardHolderName,
+                CardNumber = rentalResponseModel.CardNumber,
+                Cvc = rentalResponseModel.Cvc,
+                ExpireMonth = rentalResponseModel.ExpireMonth,
+                ExpireYear = rentalResponseModel.ExpireYear
+            };
+
+            var rental = new Rental()
+            {
+                CarId = rentalResponseModel.CarId,
+                CustomerId = 0,
+                RentDate = rentalResponseModel.FromDate,
+                ReturnDate = rentalResponseModel.ToDate,
+            };
+
+            var paymentResult = _paymentEngine.Insert(payment);
+            if (!paymentResult.Success)
+            {
+                return new ErrorDataResult<Rental>(message: paymentResult.Message);
+            }
+            var rentalResult = Insert(rental);
+
+            if (!rentalResult.Success)
+            {
+                return rentalResult;
+            }
+
+            _paymentEngine.InsertMap(new PaymentRentalMap()
+            {
+                PaymentId = paymentResult.Data.Id,
+                RentalId = rentalResult.Data.Id,
+                Price = rentalResponseModel.Price
+            });
+
+            return rentalResult;
         }
 
         public IDataResult<Rental> Update(Rental rental)
